refactor GacPage frontend
