+++ client/src/pages/css/GacPage.module.css
@@ -14,6 +14,15 @@
     flex-direction: row;
   }
 
+  .badgeDiv {
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: row;
+    justify-content: center;
+    align-content: center;
+    align-items: center;
+  }
+
   .buttonsDiv {
     display: flex;
     flex-wrap: wrap;
@@ -53,6 +62,12 @@
   }
 }
 
+.messageDiv {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
 .modal85 {
   width: 85vw;
   max-width: 85vw;
@@ -68,7 +83,7 @@
 
 .principalTable {
   place-self: center;
-  margin: 1rem 6em;
+  margin-bottom: 1rem;
   display: grid;
   place-self: center;
   justify-items: stretch;
@@ -91,10 +106,8 @@
 }
 
 .principalBody {
-  margin: 1rem 6em;
   display: grid;
-  text-align: center;
-  justify-content: center;
+  text-align: left;
 }
 
 .principalBody .principalBody  {
@@ -102,10 +115,12 @@
 }
 
 .principalBody Button {
-  margin: 0 10px;
-  background: #248be3;
-  border-color: #248be3;
+  margin-left: 10px;
   color: white;
+}
+
+.principalBody Button:first-child {
+  margin: 0;
 }
 
 .ControlDisable {
@@ -150,6 +165,7 @@
   align-content: center;
   flex-wrap: nowrap;
   height: 2.15em;
+  margin-bottom: 1rem;
 }
 
 
@@ -205,7 +221,6 @@
 }
 
 .allNonActiveCard {
-  margin-top: .5em;
   gap: 0.5em;
   width: 100%;
   height: auto;
@@ -440,4 +455,14 @@
   color: transparent !important;
   background-color: transparent !important;
   border-color: transparent !important;
+}
+
+.tableClass {
+  border-collapse: collapse;
+}
+
+.tableClass > td {
+  border: 0px none;
+  border-top: none !important;
+  border-bottom: none !important;
 }
